Split rpcjupiter and more updates
