+++ src/makefile.unix
@@ -309,6 +309,7 @@ OBJS= \
     obj/rpcwallet.o \
     obj/rpcblockchain.o \
     obj/rpcrawtransaction.o \
+	obj/rpcjupiter.o \
     obj/rpcsmessage.o \
     obj/script.o \
     obj/sync.o \
@@ -367,6 +368,7 @@ OBJS= \
     obj/rpcblockchain.o \
     obj/rpcrawtransaction.o \
     obj/rpcsmessage.o \
+	obj/rpcjupiter.o \
     obj/script.o \
     obj/sync.o \
     obj/util.o \
